Mod: 로그인 API 응답 처리 로직을 UI로 이동
